Add more explanation of why junior members
Since chatting about this effort with some more senior members of the community, it has become clear that as a person gets more senior, they start the think the community is "getting better" because they are experiencing less abuse due to their seniority, not because of actual improvements. Continuously staffing this network with junior members will be key to making sure we are getting a complete picture of the community and the effects of our efforts.